Separate out the map content.
